Create MicrOrbitals 1st project
Make sure to replace "contacts.csv" with the desired name for your CSV file. This script initializes a 'ContactManager' class, which can then be used to interact with the contact information stored in the CSV file. You can customize and extend this script based on your specific requirements. Additionally, remember to avoid committing sensitive information, such as personal details, when sharing your code on GitHub.